requirements.txt
Django==1.8.4
django-grappelli==2.7.1
django-tagulous==0.11.1
PyMySQL==0.6.6
python-slugify==1.1.3
Unidecode==0.4.19
wheel==0.24.0
